write the inverse in torch - not tested
