+++ robustnn/plnet_torch/bilipnet.py
@@ -48,3 +48,30 @@ class BiLipNet(nn.Module):
             x = self.mon_layers[k](x)
         x = self.orth_layers[self.nlayer](x)
         return x 
+    
+    def inverse(self, y,
+                alphas: Sequence[float],
+                inverse_activation_fns: Sequence[callable],
+                iterations: Sequence[int],
+                Lambdas: Sequence[float]):
+
+        """        Inverse of the BiLipNet.
+        Args:
+            y (torch.Tensor): Input tensor to be inverted.
+            alphas (Sequence[float]): Sequence of alpha values for each layer.
+            inverse_activation_fns (Sequence[callable]): Sequence of inverse activation functions for each layer.
+            iterations (Sequence[int]): Number of iterations for each layer's solver.
+            Lambdas (Sequence[float]): Step sizes for each layer's solver.
+        Returns:
+            torch.Tensor: Inverted tensor.
+        """
+        for k in range(self.depth, 0, -1):
+            x = self.orth_layers[k].inverse(y)
+            x = self.mon_layers[k-1].inverse(
+                x, 
+                alpha=alphas[k-1],
+                inverse_activation_fn=inverse_activation_fns[k-1],
+                iterations=iterations[k-1],
+                Lambda=Lambdas[k-1])
+        x = self.orth_layers[0].inverse( x)
+        return x
